refactor: create number pkg
